documind/backend/app.py
"""
DocuMind Backend API Server
Extends Pathway RAG with enterprise features
"""

import logging
from contextlib import asynccontextmanager
from typing import Optional

import pathway as pw
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException, Depends, status, UploadFile, File
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pathway.xpacks.llm.question_answering import SummaryQuestionAnswerer
from pathway.xpacks.llm.servers import QASummaryRestServer
from pydantic import BaseModel

load_dotenv()

# Set Pathway license
pw.set_license_key("demo-license-key-with-telemetry")

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s %(name)s %(levelname)s %(message)s",
    datefmt="%Y-%m-%d %H:%M:%S",
)

logger = logging.getLogger(__name__)

# ============================================================================
# Pydantic Models
# ============================================================================

class AskRequest(BaseModel):
    question: str
    return_context_docs: bool = True
    filters: Optional[dict] = None

class AskResponse(BaseModel):
    response: str
    context_docs: Optional[list] = None

class DocumentMetadata(BaseModel):
    path: str
    name: str
    size: int
    type: str
    status: str

class RAGConfigModel(BaseModel):
    llm_model: str
    embedding_model: str
    search_topk: int = 6
    temperature: float = 0.0
    max_tokens: int = 2000

class DataSourceModel(BaseModel):
    id: Optional[str] = None
    type: str
    name: str
    config: dict
    status: str = "active"

# ============================================================================
# Authentication
# ============================================================================

security = HTTPBearer()

def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    """
    Verify JWT token (simplified for demo)
    In production, implement proper JWT verification
    """
    token = credentials.credentials
    # TODO: Implement real JWT verification
    if token != "demo-token":
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication credentials",
        )
    return token

# ============================================================================
# FastAPI Application
# ============================================================================

@asynccontextmanager
async def lifespan(app: FastAPI):
    """Manage application lifecycle"""
    logger.info("Starting DocuMind Backend API")
    # Initialize Pathway app here if needed
    yield
    logger.info("Shutting down DocuMind Backend API")

app = FastAPI(
    title="DocuMind API",
    description="Enterprise AI Document Intelligence Platform",
    version="0.1.0",
    lifespan=lifespan,
)

# CORS Configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:8501"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================================================
# Health & Status Endpoints
# ============================================================================

@app.get("/")
async def root():
    return {
        "name": "DocuMind API",
        "version": "0.1.0",
        "status": "running"
    }

@app.get("/health")
async def health():
    return {"status": "healthy"}

# ============================================================================
# Document Management Endpoints
# ============================================================================

@app.post("/v1/documents/upload")
async def upload_document(
    file: UploadFile = File(...),
    token: str = Depends(verify_token)
):
    """Upload a new document for indexing"""
    try:
        # TODO: Implement file upload to storage
        # For now, return mock response
        return {
            "id": "doc_123",
            "name": file.filename,
            "size": file.size,
            "status": "processing",
            "message": "Document uploaded successfully"
        }
    except Exception as e:
        logger.error(f"Error uploading document: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.delete("/v1/documents/{document_id}")
async def delete_document(
    document_id: str,
    token: str = Depends(verify_token)
):
    """Delete a document"""
    # TODO: Implement document deletion
    return {"message": f"Document {document_id} deleted successfully"}

# ============================================================================
# RAG Configuration Endpoints
# ============================================================================

@app.get("/v1/config/rag", response_model=RAGConfigModel)
async def get_rag_config(token: str = Depends(verify_token)):
    """Get current RAG configuration"""
    # TODO: Load from config file or database
    return RAGConfigModel(
        llm_model="gpt-4.1-mini",
        embedding_model="text-embedding-3-small",
        search_topk=6,
        temperature=0.0,
        max_tokens=2000,
    )

@app.put("/v1/config/rag", response_model=RAGConfigModel)
async def update_rag_config(
    config: RAGConfigModel,
    token: str = Depends(verify_token)
):
    """Update RAG configuration"""
    # TODO: Save to config file or database
    logger.info(f"Updating RAG config: {config}")
    return config

# ============================================================================
# Data Source Management Endpoints
# ============================================================================

@app.get("/v1/datasources")
async def list_datasources(token: str = Depends(verify_token)):
    """List all configured data sources"""
    # TODO: Load from database
    return [
        {
            "id": "ds_1",
            "type": "local",
            "name": "Local Documents",
            "status": "active",
            "lastSync": "2025-01-15T10:30:00Z"
        }
    ]

@app.post("/v1/datasources")
async def create_datasource(
    source: DataSourceModel,
    token: str = Depends(verify_token)
):
    """Create a new data source"""
    # TODO: Save to database and update app.yaml
    logger.info(f"Creating data source: {source}")
    return {
        "id": "ds_new",
        "message": "Data source created successfully",
        **source.dict()
    }

@app.post("/v1/datasources/{source_id}/sync")
async def sync_datasource(
    source_id: str,
    token: str = Depends(verify_token)
):
    """Trigger sync for a data source"""
    # TODO: Implement sync logic
    return {"message": f"Sync triggered for data source {source_id}"}

# ============================================================================
# Statistics Endpoint
# ============================================================================

@app.get("/v1/stats")
async def get_stats(token: str = Depends(verify_token)):
    """Get system statistics"""
    return {
        "total_documents": 1234,
        "indexed_documents": 1180,
        "questions_this_month": 567,
        "storage_used_gb": 45.2,
        "storage_limit_gb": 100.0,
    }

# ============================================================================
# Authentication Endpoints (Mock)
# ============================================================================

class LoginRequest(BaseModel):
    email: str
    password: str

class RegisterRequest(BaseModel):
    email: str
    password: str
    name: str
    organization: str

@app.post("/v1/auth/login")
async def login(request: LoginRequest):
    """User login"""
    # TODO: Implement real authentication
    if request.email and request.password:
        return {
            "token": "demo-token",
            "user": {
                "id": "user_1",
                "email": request.email,
                "name": "Demo User",
                "role": "admin"
            }
        }
    raise HTTPException(status_code=401, detail="Invalid credentials")

@app.post("/v1/auth/register")
async def register(request: RegisterRequest):
    """User registration"""
    # TODO: Implement real registration
    return {
        "message": "User registered successfully",
        "user": {
            "id": "user_new",
            "email": request.email,
            "name": request.name
        }
    }

# ============================================================================
# Main Entry Point
# ============================================================================

if __name__ == "__main__":
    import uvicorn

    # Load Pathway RAG app configuration
    with open("app.yaml") as f:
        pathway_config = pw.load_yaml(f)

    # Start Pathway in background
    # TODO: Integrate Pathway app properly

    # Start FastAPI server
    uvicorn.run(
        app,
        host="0.0.0.0",
        port=8000,
        log_level="info"
    )
